% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{bcg_app}
\alias{bcg_app}
\title{Launch BCG annotation app}
\usage{
bcg_app(...)
}
\arguments{
\item{...}{Additional parameters to shinyApp}
}
\value{
Runs a Shiny app
}
\description{
Launch BCG annotation app
}
