% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bcg.R
\name{plot_bcg}
\alias{plot_bcg}
\title{Plot BCG}
\usage{
plot_bcg(prh, beats)
}
\arguments{
\item{prh}{PRH in tibble format (see read_nc())}

\item{beats}{Subset of PRH rows corresponding to heart beats}
}
\value{
plotly object
}
\description{
Creates an interactive plotly object with depth and jerk profiles.
}
