% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom_beats}
\alias{zoom_beats}
\title{Zoom into heart beats}
\usage{
zoom_beats(beats, range = NULL)
}
\arguments{
\item{beats}{Heart beats tibble (subset of PRH)}

\item{range}{Time range of zoomed region (POSIXct vector length 2)}
}
\value{
Zoomed beats tibble
}
\description{
Zoom into heart beats
}
