% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom_prh}
\alias{zoom_prh}
\title{Zoom into PRH}
\usage{
zoom_prh(prh, range = NULL, maxpts = 10000)
}
\arguments{
\item{prh}{PRH in tibble format (see read_nc())}

\item{range}{Time range of zoomed region (POSIXct vector length 2)}

\item{maxpts}{Maximum number of points to decimate to (10,000 by default)}
}
\value{
Zoomed and decimated PRH
}
\description{
Zoom into PRH
}
